Updated map layer to be more visually appealing
This small change brought about 2 weird browser issues. Apparently Chrome is warning developers that their third party cookie restrictions are about to get stricter, so they are warning us that our code may not function properly in the near future. This change is not supposed to happen until July at the earliest, so I think it is safe to ignore for the scope of this class.

I also changed the map height to fit my screen.